+++ loanapplicationform/src/main/java/in/srb/service/LoanServiceI.java
@@ -27,7 +27,7 @@ public interface LoanServiceI {
 
 	public Customer changestatus(int customerId);
 
-	public Customer createleager(int customerId, int payment) throws Exception;
+	public Customer createleager(int customerId, Double payment) throws Exception;
